+++ src/providers/strategies/CrossWindowProviderStrategy/index.ts
@@ -1,2 +1,1 @@
 export * from './CrossWindowProviderStrategy';
-export * from './types';